# [ENH] Faster normalization