--- a/Orange/preprocess/normalize.py
+++ b/Orange/preprocess/normalize.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from Orange.data import Domain, ContinuousVariable
-from Orange.statistics import distribution
+from Orange.statistics import basic_stats
 from Orange.util import Reprable
 from .preprocess import Normalize
 from .transformation import Normalizer as Norm
@@ -22,30 +22,34 @@ class Normalizer(Reprable):
         self.normalize_datetime = normalize_datetime
 
     def __call__(self, data):
-        dists = distribution.get_distributions(data)
-        new_attrs = [self.normalize(dists[i], var) for
+        stats = basic_stats.DomainBasicStats(data, compute_variance=True)
+        new_attrs = [self.normalize(stats[i], var) for
                      (i, var) in enumerate(data.domain.attributes)]
 
         new_class_vars = data.domain.class_vars
         if self.transform_class:
             attr_len = len(data.domain.attributes)
-            new_class_vars = [self.normalize(dists[i + attr_len], var) for
+            new_class_vars = [self.normalize(stats[i + attr_len], var) for
                               (i, var) in enumerate(data.domain.class_vars)]
 
         domain = Domain(new_attrs, new_class_vars, data.domain.metas)
         return data.transform(domain)
 
-    def normalize(self, dist, var):
+    def normalize(self, stats, var):
         if not var.is_continuous or (var.is_time and not self.normalize_datetime):
             return var
         elif self.norm_type == Normalize.NormalizeBySD:
-            var = self.normalize_by_sd(dist, var)
+            var = self.normalize_by_sd(stats, var)
         elif self.norm_type == Normalize.NormalizeBySpan:
-            var = self.normalize_by_span(dist, var)
+            var = self.normalize_by_span(stats, var)
         return var
 
-    def normalize_by_sd(self, dist, var: ContinuousVariable) -> ContinuousVariable:
-        avg, sd = (dist.mean(), dist.standard_deviation()) if dist.size else (0, 1)
+    def normalize_by_sd(self, stats, var: ContinuousVariable) -> ContinuousVariable:
+        avg, sd = (stats.mean, stats.var**0.5)
+        if np.isnan(avg):
+            avg = 0
+        if np.isnan(sd):
+            sd = 1
         if sd == 0:
             sd = 1
         if self.center:
@@ -62,8 +66,8 @@ class Normalizer(Reprable):
 
         return var.copy(compute_value=compute_val, number_of_decimals=num_decimals)
 
-    def normalize_by_span(self, dist, var: ContinuousVariable) -> ContinuousVariable:
-        dma, dmi = (dist.max(), dist.min()) if dist.shape[1] else (np.nan, np.nan)
+    def normalize_by_span(self, stats, var: ContinuousVariable) -> ContinuousVariable:
+        dma, dmi = (stats.max, stats.min)
         diff = dma - dmi
         if diff < 1e-15:
             diff = 1