##### Issue
Fixes #5219

##### Description of changes
The computation of distributions is slow (it works through contingencies and requires sorting).

This slowdown was especially problematic for dask tables.

~~The new code does not currently support weights! Supporting them would require making a weighted variance computation. Should we go there?~~ Done in #6205.

Requires #6204 and #6205.

Testing performance on 1000x1000 tables.

Before:
```
[normalize_only_parameters] with 5 loops, best of 3:
	min 233 msec per loop
	avg 235 msec per loop
[normalize_only_transform] with 5 loops, best of 3:
	min 120 msec per loop
	avg 121 msec per loop
```

After:
```
[normalize_only_parameters] with 5 loops, best of 3:
	min 71.7 msec per loop
	avg 73.3 msec per loop
[normalize_only_transform] with 5 loops, best of 3:
	min 117 msec per loop
	avg 119 msec per loop
```

The mean/variance estimation part is 3x faster. Of course, this PR has no influence on the actual transformation (the domain contains the same variables).

##### Includes
- [X] Code changes
- [X] Tests
- [ ] Documentation